requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
slack_bolt==1.19.1
slack_sdk==3.33.4
httpx==0.27.2
sqlite-utils==3.37
aiosqlite==0.20.0
typer==0.12.5
python-dateutil==2.9.0.post0
orjson==3.10.7
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
langchain==0.2.14
langgraph==0.2.18
langchain-openai==0.1.22
langchain-groq==0.1.8
jinja2==3.1.4
authlib==1.3.1
tavily-python==0.3.3
rapidfuzz==3.9.6
python-jose==3.3.0